Move cursor capture to separate file
